`tyger auth` command
