+++ cli/cmd/tyger/main.go
@@ -56,10 +56,10 @@ func newRootCommand() *cobra.Command {
 	rootCommand.AddCommand(cmd.NewBufferCommand())
 	rootCommand.AddCommand(cmd.NewCodespecCommand())
 	rootCommand.AddCommand(cmd.NewRunCommand())
+	rootCommand.AddCommand(install.NewAuthCommand())
 	rootCommand.AddCommand(install.NewConfigCommand(rootCommand))
 	rootCommand.AddCommand(install.NewCloudCommand(rootCommand))
 	rootCommand.AddCommand(install.NewApiCommand(rootCommand))
-	rootCommand.AddCommand(install.NewIdentitiesCommand(rootCommand))
 	rootCommand.AddCommand(cmd.NewStdioProxyCommand())
 
 	var setSortFlagsFalse func(*cobra.Command)
