+++ cli/internal/controlplane/model/model.go
@@ -15,7 +15,10 @@ import (
 type ServiceMetadata struct {
 	Authority      string   `json:"authority,omitempty"`
 	Audience       string   `json:"audience,omitempty"`
+	ApiAppUri      string   `json:"serverAppUri,omitempty"`
+	ApiAppId       string   `json:"serverAppId,omitempty"`
 	CliAppUri      string   `json:"cliAppUri,omitempty"`
+	CliAppId       string   `json:"cliAppId,omitempty"`
 	DataPlaneProxy string   `json:"dataPlaneProxy,omitempty"`
 	Capabilities   []string `json:"capabilities,omitempty"`
 	ApiVersions    []string `json:"apiVersions,omitempty"`
