+++ cli/internal/install/cloudinstall/config-pretty.tpl
@@ -242,6 +242,9 @@ organizations:
         tenantId: {{ .Auth.TenantID }}
         apiAppUri: {{ .Auth.ApiAppUri }}
         cliAppUri: {{ .Auth.CliAppUri }}
+        {{- if .ServiceManagementReference }}
+        serviceManagementReference: {{ .ServiceManagementReference }}
+        {{- end }}
 
       {{- if (and .Buffers (or .Buffers.ActiveLifetime .Buffers.SoftDeletedLifetime)) }}
